% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{crant_cave_client}
\alias{crant_cave_client}
\title{Low level access to crant's CAVE annotation infrastructure}
\usage{
crant_cave_client()
}
\value{
A reticulate R object wrapping the python CAVEclient.
}
\description{
Low level access to crant's CAVE annotation infrastructure
}
\examples{
\dontrun{
fcc=crant_cave_client()
tables=fcc$annotation$get_tables()
fcc$materialize$get_table_metadata(tables[1])
}
}
