% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{crant_cave_tables}
\alias{crant_cave_tables}
\title{Read CRANT CAVE-tables, good sources of metadata}
\usage{
crant_cave_tables(datastack_name = "kronauer_ant", select = NULL, ...)
}
\arguments{
\item{datastack_name}{Defaults to "kronauer_ant" i.e. CRANTb.}

\item{select}{A regex term for the name of the table you want}

\item{...}{Additional arguments passed to
\code{fafbseg::\link{flywire_cave_query}}}
}
\value{
A \code{data.frame} describing a CAVE-table related to the CRANT project.
In the case of \code{crant_cave_tables}, a vector is returned containing the names of
all query-able cave tables.
}
\description{
Read CRANT CAVE-tables, good sources of metadata
}
\examples{
\dontrun{
all_crant_soma_positions <- crant_nuclei()
points3d(nat::xyzmatrix(all_crant_soma_positions$pt_position))

# Another way to query a specific table
crant_backbone_proofread <- with_crant(bancr:::get_cave_table_data("backbone_proofread"))
}
}
\seealso{
\code{fafbseg::\link{flywire_cave_query}}
}
