% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{crant_islatest}
\alias{crant_islatest}
\title{Check if a CRANT root id is up to date}
\usage{
crant_islatest(x, timestamp = NULL, ...)
}
\arguments{
\item{x}{FlyWire rootids in any format understandable to
\code{\link[fafbseg]{ngl_segments}} including as \code{integer64}}

\item{timestamp}{(optional) argument to set an endpoint - edits after this
time will be ignored (see details).}

\item{...}{Additional arguments passed to \code{\link{flywire_islatest}}}
}
\description{
Check if a CRANT root id is up to date
}
\examples{
\dontrun{
crant_islatest("576460752684030043")
}
}
\seealso{
Other crant-ids: 
\code{\link{crant_latestid}()},
\code{\link{crant_leaves}()},
\code{\link{crant_rootid}()},
\code{\link{crant_xyz2id}()}
}
\concept{crant-ids}
