% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{crant_latestid}
\alias{crant_latestid}
\alias{crant_updateids}
\title{Find the latest id for a CRANT root id}
\usage{
crant_latestid(
  rootid,
  sample = 1000L,
  cloudvolume.url = NULL,
  Verbose = FALSE,
  ...
)

crant_updateids(
  x,
  root.column = "root_id",
  supervoxel.column = "supervoxel_id",
  position.column = "position",
  ...
)
}
\arguments{
\item{rootid}{One ore more FlyWire rootids defining a segment (in any form
interpretable by \code{\link[fafbseg]{ngl_segments}})}

\item{sample}{An absolute or fractional number of supervoxel ids to map to
rootids or \code{FALSE} (see details).}

\item{cloudvolume.url}{URL for CloudVolume to fetch segmentation image data.
The default value of NULL chooses the flywire production segmentation
dataset.}

\item{Verbose}{When set to \code{TRUE} prints information about what fraction
of}

\item{...}{Additional arguments passed to \code{\link{flywire_latestid}}}

\item{x}{a \code{data.frame} with at least one of: \code{root_id}, \code{pt_root_id}, \code{supervoxel_id} and/or \code{pt_supervoxel_id}.
Supervoxels will be preferentially used to update the \code{root_id} column.
Else a vector of \code{CRANT} root IDs.}

\item{root.column}{when \code{x} is a \code{data.frame}, the \code{root_id} column you wish to update}

\item{supervoxel.column}{when \code{x} is a \code{data.frame}, the \code{supervoxel_id} column you wish to use to update \code{root.column}}

\item{position.column}{when \code{x} is a \code{data.frame}, the \code{position} column with xyz values you wish to use to update \code{supervoxel.column}}
}
\description{
Find the latest id for a CRANT root id
}
\examples{
\dontrun{
crant_latestid("576460752684030043")
}
}
\seealso{
\code{\link{crant_islatest}}

Other crant-ids: 
\code{\link{crant_islatest}()},
\code{\link{crant_leaves}()},
\code{\link{crant_rootid}()},
\code{\link{crant_xyz2id}()}
}
\concept{crant-ids}
