% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{crant_leaves}
\alias{crant_leaves}
\title{Find the supervoxel identifiers of a CRANT neuron}
\usage{
crant_leaves(x, integer64 = TRUE, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{integer64}{Whether to return ids as integer64 type (the default, more
compact but a little fragile) rather than character (when \code{FALSE}).}

\item{...}{additional arguments passed to \code{\link{flywire_leaves}}}
}
\value{
A vector of supervoxel ids
}
\description{
Find the supervoxel identifiers of a CRANT neuron
}
\examples{
\dontrun{
svids=crant_leaves("576460752684030043")
head(svids)
}
}
\seealso{
\code{\link{flywire_leaves}}

Other crant-ids: 
\code{\link{crant_islatest}()},
\code{\link{crant_latestid}()},
\code{\link{crant_rootid}()},
\code{\link{crant_xyz2id}()}
}
\concept{crant-ids}
