% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{crant_partner_summary}
\alias{crant_partner_summary}
\alias{crant_partners}
\title{Summarise the connectivity of CRANTb neurons}
\usage{
crant_partner_summary(
  rootids,
  partners = c("outputs", "inputs"),
  synapse_table = c("synapses_v2"),
  threshold = 0,
  remove_autapses = TRUE,
  cleft.threshold = 0,
  datastack_name = NULL,
  ...
)

crant_partners(
  rootids,
  partners = c("input", "output"),
  synapse_table = c("synapses_v2"),
  ...
)
}
\arguments{
\item{rootids}{Character vector specifying one or more crant rootids. As a
convenience this argument is passed to \code{\link{crant_ids}} allowing you
to pass in data.frames, crant URLs or simple ids.}

\item{partners}{Character vector, either "outputs" or "inputs" to specify the direction of synaptic connections to retrieve.}

\item{synapse_table}{Character, the name of the synapse CAVE table you wish to use. Defaults to the latest.}

\item{threshold}{Integer threshold for minimum number of synapses (default 0).}

\item{remove_autapses}{Logical, whether to remove self-connections (default TRUE).}

\item{cleft.threshold}{Numeric threshold for cleft filtering (default 0).}

\item{datastack_name}{An optional CAVE \code{datastack_name}. If unset a
sensible default is chosen.}

\item{...}{Additional arguments passed to \code{\link[fafbseg]{flywire_partner_summary}}}
}
\value{
a data.frame
}
\description{
Returns synaptically connected partners for specified neurons. Understanding
synaptic partnerships is crucial for analyzing neural circuits in the CRANTb
connectome.

\code{crant_partners} returns details of each unitary synaptic
connection (including its xyz location).
}
\details{
note that the rootids you pass in must be up to date. See example.
}
\examples{
\dontrun{
# Basic connectivity analysis
sample_id=crant_latestid("576460752716912866")
head(crant_partner_summary(sample_id))
head(crant_partner_summary(sample_id, partners='inputs'))
}
\dontrun{
# plot input and output synapses of a neuron
nclear3d()
fpi=crant_partners(crant_latestid("576460752716912866"), partners='in')
points3d(crant_raw2nm(fpi$post_pt_position), col='cyan')
fpo=crant_partners(crant_latestid("576460752716912866"), partners='out')
points3d(crant_raw2nm(fpo$pre_pt_position), col='red')
}
}
\seealso{
\code{\link{flywire_partner_summary}}, \code{\link{crant_latestid}}
}
