% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{crant_read_neuroglancer_mesh}
\alias{crant_read_neuroglancer_mesh}
\title{Read CRANT neuroglancer meshes, e.g., ROI meshes}
\usage{
crant_read_neuroglancer_mesh(
  x = 1,
  url = paste0("https://www.googleapis.com/storage/v1/b/",
    "dkronauer-ant-001-alignment-final/o/tissue_mesh\%2F",
    "mesh\%2Ftissue_mesh.frag?alt=media",
    "&neuroglancer=a2b0cf07baf8c501891d6c683cc7e24a"),
  ...
)
}
\arguments{
\item{x}{the numeric identifier that specifies the mesh to read, defaults to \code{1} the CRANTb outline mesh.}

\item{url}{the URL that directs \code{bancr} to where CRANT meshes are stored.}

\item{...}{additional arguments to \code{\link{GET}}}
}
\value{
a mesh3d object for the specified mesh.
}
\description{
Read CRANT neuroglancer meshes, e.g., ROI meshes
}
\examples{
\dontrun{
# default is brain mesh
crant.mesh  <- crant_read_neuroglancer_mesh()
}
}
\seealso{
\code{\link{crant_read_neuron_meshes}}
}
