% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{crant_rootid}
\alias{crant_rootid}
\title{Find the root identifier of a CRANT neuron}
\usage{
crant_rootid(x, integer64 = FALSE, ...)
}
\arguments{
\item{x}{One or more FlyWire segment ids}

\item{integer64}{Whether to return ids as integer64 type (more compact but a
little fragile) rather than character (default \code{FALSE}).}

\item{...}{Additional arguments passed to \code{pbapply::pbsapply} and
eventually to Python \code{cv$CloudVolume} object.}
}
\value{
A vector of root ids (by default character)
}
\description{
Find the root identifier of a CRANT neuron
}
\examples{
\dontrun{
crant_rootid("576460752684030043")
}
}
\seealso{
\code{\link{flywire_rootid}}

Other crant-ids: 
\code{\link{crant_islatest}()},
\code{\link{crant_latestid}()},
\code{\link{crant_leaves}()},
\code{\link{crant_xyz2id}()}
}
\concept{crant-ids}
