% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crant.R
\name{crant_set_token}
\alias{crant_set_token}
\title{Set the token to be used to authenticate to CRANT autosegmentation resources}
\usage{
crant_set_token(token = NULL, domain = "https://proofreading.zetta.ai")
}
\arguments{
\item{token}{An optional token string. When missing you are prompted to
generate a new token via your browser.}

\item{domain}{the domain for which your CAVE token is valid, i.e. where the
project is hosted.}
}
\value{
The path to the token file (invisibly)
}
\description{
Set the token to be used to authenticate to CRANT autosegmentation resources
}
