% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{crant_xyz2id}
\alias{crant_xyz2id}
\title{Convert xyz locations to CRANT root or supervoxel ids}
\usage{
crant_xyz2id(
  xyz,
  rawcoords = FALSE,
  root = TRUE,
  integer64 = FALSE,
  fast_root = TRUE,
  method = c("cloudvolume", "spine"),
  ...
)
}
\arguments{
\item{xyz}{One or more xyz locations as an Nx3 matrix or in any form
compatible with \code{\link[fafbseg]{xyzmatrix}} including \code{neuron} or
\code{mesh3d} surface objects.}

\item{rawcoords}{whether the input values are raw voxel indices or in nm}

\item{root}{Whether to return the root id of the whole segment rather than
the supervoxel id.}

\item{integer64}{Whether to return ids as integer64 type (more compact but a
little fragile) rather than character (default \code{FALSE}).}

\item{fast_root}{Whether to use a fast but two-step look-up procedure when
finding roots. This is strongly recommended and the alternative approach
has only been retained for validation purposes.}

\item{method}{Whether to use the
\href{https://services.itanna.io/app/transform-service/docs}{spine
transform-service} API or cloudvolume for lookup. \code{"auto"} is
presently a synonym for \code{"spine"}.}

\item{...}{additional arguments passed to \code{pbapply} when looking up
multiple positions.}
}
\value{
A character vector of segment ids, NA when lookup fails.
}
\description{
Convert xyz locations to CRANT root or supervoxel ids
}
\details{
This used to be very slow because we do not have a supervoxel
field on spine.

I am somewhat puzzled by the voxel dimensions for crant. Neuroglancer
clearly shows voxel coordinates of 4.3x4.3x45. But in this function, the
voxel coordinates must be set to 4.25 in x-y to give the correct answers.
}
\examples{
\dontrun{
# a point from neuroglancer, should map to 576460752684030043
crant_xyz2id(cbind(37306, 31317, 1405), rawcoords=TRUE)
}
}
\seealso{
\code{\link{flywire_xyz2id}}

Other crant-ids: 
\code{\link{crant_islatest}()},
\code{\link{crant_latestid}()},
\code{\link{crant_leaves}()},
\code{\link{crant_rootid}()}
}
\concept{crant-ids}
