% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crantb.surf}
\alias{crantb.surf}
\title{Simplified tissue surface for crant}
\format{
An object of class \code{hxsurf} (inherits from \code{list}) of length 4.
}
\usage{
crantb.surf
}
\description{
\code{crantb.surf} is unsymmetrical and not a normalised version of the dataset mesh.
It is the outline of the dataset in nanometers. It can be seen in neuroglancer
\href{https://spelunker.cave-explorer.org/#!middleauth+https://global.daf-apis.com/nglstate/api/v1/5185996221054976}{here}.
}
\examples{
\dontrun{
# Depends on nat
library(nat)
rgl::wire3d(crantb.surf)
}
}
\keyword{datasets}
