% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crant-package.R
\docType{package}
\name{crantr-package}
\alias{crantr}
\alias{crantr-package}
\title{crantr: R Client Access to Clonal Raider ANT (CRANT) Datasets}
\description{
Access to the Clonal Raider ANT (CRANT) EM datasets, starting with CRANTb and focussing on support for autosegmentation. Organised as a wrapper around the fafbseg and bancr packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flyconnectome/crantr}
  \item \url{https://flyconnectome.github.io/crantr/}
  \item Report bugs at \url{https://github.com/flyconnectome/crantr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

}
\keyword{internal}
