% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crant.R
\name{choose_crant}
\alias{choose_crant}
\alias{with_crant}
\title{Choose or (temporarily) use the CRANT autosegmentation}
\usage{
choose_crant(set = TRUE)

with_crant(expr)
}
\arguments{
\item{set}{Whether or not to permanently set the CRANT auto-segmentation as the
default for \code{\link{fafbseg}} functions.}

\item{expr}{An expression to evaluate while CRANT is the default
autosegmentation}
}
\value{
If \code{set=TRUE} a list containing the previous values of the
relevant global options (in the style of \code{\link{options}}. If
\code{set=FALSE} a named list containing the option values.
}
\description{
Choose or (temporarily) use the CRANT autosegmentation
}
\details{
\code{bancr} inherits a significant amount of infrastructure from
the \code{\link{fafbseg}} package. This has the concept of the
\emph{active} autosegmentation, which in turn defines one or more R options
containing URLs pointing to voxel-wise segmentation, mesh etc data. These
are normally contained within a single neuroglancer URL which points to
multiple data layers. For banc this is the neuroglancer scene returned by
\code{\link{crant_scene}}.
}
\examples{
\dontrun{
choose_crant()
options()[grep("^fafbseg.*url", names(options()))]
}
\dontrun{
with_crant(fafbseg::flywire_islatest('576460752653449509'))
}
\dontrun{
with_crant(fafbseg::flywire_latestid('576460752653449509'))
}
}
