% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cave-tables.R
\name{crant_cave_query}
\alias{crant_cave_query}
\alias{crant_backbone_proofread}
\title{Query CRANT tables in the CAVE annotation system}
\usage{
crant_cave_query(table, live = 2, ...)

crant_backbone_proofread(...)
}
\arguments{
\item{table}{The name of the table (or view, see views section) to query}

\item{live}{Whether to use live query mode, which updates any root ids to
their current value (or to another \code{timestamp} when provided). Values
of \code{TRUE} or \code{1} select CAVE's \emph{Live} mode, while \code{2}
selects \code{Live live} mode which gives access even to annotations that
are not part of a materialisation version. See section \bold{Live and Live
Live queries} for details.}

\item{...}{Additional arguments passed to \code{\link{flywire_cave_query}}
\code{\link[fafbseg]{flywire_cave_query}}}
}
\value{
A data.frame
}
\description{
Query CRANT tables in the CAVE annotation system
}
\examples{
\dontrun{
backbone.proofread=crant_cave_query('backbone_proofread')
View(backbone.proofread)
}
}
\seealso{
\code{\link[fafbseg]{flywire_cave_query}}
}
\concept{crant-cave}
