% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crant.R
\name{crant_ids}
\alias{crant_ids}
\title{Make sure given root IDs look like CRANT root IDs}
\usage{
crant_ids(x, integer64 = NA)
}
\arguments{
\item{x}{A data.frame, URL or vector of ids}

\item{integer64}{Whether to return ids as \code{bit64::integer64} or
character vectors. Default value of NA leaves the ids unmodified.}
}
\description{
Make sure given root IDs look like CRANT root IDs
}
