% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconat.R
\name{crant_meta}
\alias{crant_meta}
\title{Query cached CRANTb meta data}
\usage{
crant_meta(ids = NULL)
}
\arguments{
\item{ids}{Vector of neuron/root IDs to select, or \code{NULL} for all.}
}
\value{
tibble/data.frame, possibly filtered by ids.
}
\description{
Returns results from the in-memory cache, filtered by \code{ids} if given.
Cache must be created first using \code{\link[=crant_meta_create_cache]{crant_meta_create_cache()}}.
}
\details{
\code{crant_meta()} never queries databases directly.
If \code{ids} are given, filters the meta table by root_id.
}
\examples{
\dontrun{
crant_meta_create_cache() # build the cache
all_meta <- crant_meta()  # retrieve all
}
}
\seealso{
\code{\link[=crant_meta_create_cache]{crant_meta_create_cache()}}
}
