% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconat.R
\name{register_crant_coconat}
\alias{register_crant_coconat}
\title{Use CRANTb data with coconat for connectivity similarity}
\usage{
register_crant_coconat(showerror = TRUE)
}
\arguments{
\item{showerror}{Logically, error-out silently or not.}
}
\description{
Register the CRANTb dataset with \href{https://github.com/natverse/coconat}{coconat},
a natverse R package for between and within dataset connectivity comparisons using cosine similarity.
}
\details{
\code{register_crant_coconat()} registers \code{crantr}-backed functionality for use with
}
\examples{
\dontrun{
library(coconatfly)
library(bancr)
crant_meta_create_cache(use_seatable=TRUE)
banc_meta_create_cache(use_seatable=TRUE)
register_crant_coconat()
register_banc_coconat()
cf_cosine_plot(cf_ids('/type:NSC', datasets = c("crant", "banc")))
}
}
\seealso{
\code{\link[=crant_meta_create_cache]{crant_meta_create_cache()}}
}
